fix: fix sign compare error

+++ modules/planning/open_space/trajectory_smoother/distance_approach_ipopt_cuda_interface.cc
@@ -21,6 +21,7 @@
 
 #include <algorithm>
 #include <limits>
+#include <vector>
 
 #if USE_GPU == 1
 #include "modules/planning/open_space/trajectory_smoother/planning_block.h"
@@ -51,10 +52,10 @@ DistanceApproachIPOPTCUDAInterface::DistanceApproachIPOPTCUDAInterface(
       obstacles_edges_num_(obstacles_edges_num),
       obstacles_A_(obstacles_A),
       obstacles_b_(obstacles_b) {
-  ACHECK(horizon < std::numeric_limits<int>::max())
+  ACHECK(horizon < static_cast<size_t>(std::numeric_limits<int>::max()))
       << "Invalid cast on horizon in open space planner";
   horizon_ = static_cast<int>(horizon);
-  ACHECK(obstacles_num < std::numeric_limits<int>::max())
+  ACHECK(obstacles_num < static_cast<size_t>(std::numeric_limits<int>::max()))
       << "Invalid cast on obstacles_num in open space planner";
 
   obstacles_num_ = static_cast<int>(obstacles_num);
