+++ modules/planning/open_space/trajectory_smoother/distance_approach_ipopt_relax_end_slack_interface.cc
@@ -18,6 +18,10 @@
  * @file
  */
 #include "modules/planning/open_space/trajectory_smoother/distance_approach_ipopt_relax_end_slack_interface.h"
+
+#include <algorithm>
+#include <vector>
+
 // #define ADEBUG AERROR
 namespace apollo {
 namespace planning {
@@ -47,10 +51,10 @@ DistanceApproachIPOPTRelaxEndSlackInterface::
       obstacles_edges_num_(obstacles_edges_num),
       obstacles_A_(obstacles_A),
       obstacles_b_(obstacles_b) {
-  ACHECK(horizon < std::numeric_limits<int>::max())
+  ACHECK(horizon < static_cast<size_t>(std::numeric_limits<int>::max()))
       << "Invalid cast on horizon in open space planner";
   horizon_ = static_cast<int>(horizon);
-  ACHECK(obstacles_num < std::numeric_limits<int>::max())
+  ACHECK(obstacles_num < static_cast<size_t>(std::numeric_limits<int>::max()))
       << "Invalid cast on obstacles_num in open space planner";
 
   // TODO(RHe): use relax slacks from dual warm up
